==> Analyses/01_2_Data_Preparation_MarsShapPhylo.Rmd <==
---
title: "01_Mars_Shape_Phylo_Prep"
author: "Vera Weisbecker"
date: "26 June 2019"
output: html_document
---


## load required packages

```{r }
library(geomorph)
library(ape)
library(plyr)
library(abind)
library(geiger)
library(caper)
library(devtools)
library(phytools)

# My rmd file for some reason won't recognise it s local file, so this sets the wd to local
library(rstudioapi)
setwd(dirname(getActiveDocumentContext()$path))

#contains some auxiliary functions

source("Utilities.r")

```

## Read in coordinates


```{r }

filelist <- list.files(path= "../Data/Raw/Coordinates", pattern = "*.txt")


#Next step is to remove the Museum IDs from the specimen names   
  
 
  names <- gsub (".txt", "", filelist) # extracts names of specimens from the file name
  filelist <- paste("../Data/Raw/Coordinates/", filelist, sep="") # rename with path
  coords <- NULL # make empty object that will be filled with 3D array of coordinate data
  for (i in 1:length(filelist)){
    temp  <- read.morphologika(filelist[i]) 
    k <- dim(temp)[1] 
    coords <- rbind(coords, two.d.array(temp)) }
  Data <- arrayspecs(coords, k, 3) 
  dimnames(Data)[[3]] <- names
 # remove(i, filelist, names, k, coords, temp) # clean up environment
  
```


#GPA for all coordinates

```{r}

#Run GPA
GPA_AllSpecimens <- gpagen(Data, Proj = TRUE)

```

#using my Landm.place.check function to visualise each sepcimen relative to the specimen that's closest to the mean. I am just running through all specimens in batches of 10

```{r}


Landm.place.check(51,62,Meanspec)


```

#After identifying specimens with switched landmarks in the Landm.place.check function, I fixed them manually by switching coordinates in the text file (with one exception where the lms were truly out of order, and I had to open the specimen in checkpoint). The notes on which specimesn were fixed are in the "original with mistakes " folder.

```{r}
names(GPA_AllSpecimens$Csize)

plotRefToTarget(GPA_AllSpecimens$coords[,,20], GPA_AllSpecimens$coords[,,Meanspec], method = "vector", main = dimnames(GPA_AllSpecimens$coords)[[3]][20], label=TRUE,
                gridPars = gridPar(pt.size = 0.5, pt.bg = "hotpink", txt.cex = "1.5" ))

```
#Check that landmarks are read in OK and there are no outliers; also visually check that the landmarks are placed without any crossovers. 

```{r}

#these checks looks ok, with Notoryctes, Tarsipes and Vombatus predictably outliers and comfortingly duplicate specimens occupy similar spaces

PCA <- gm.prcomp(GPA_AllSpecimens$coords)
plot(PCA$x[,2]~ PCA$x[,1]  )
text(PCA$x[,2]~ PCA$x[,1], labels = dimnames(GPA_AllSpecimens$coords)[[3]])

#outliers? Again, Notoryctes and Vombatus. Just.

plotOutliers(GPA_AllSpecimens$coords)

Meanspec <- findMeanSpec(GPA_AllSpecimens$coords)


```

#Now, check that the centroid sizes are ok. Sometimes they are out by an order of magnitude, this is normally easy to see by ordering the specimens by size and checking that they match what you expect.

#I CAN SEE  A COUPLE THAT MIGHT BE WRONG. ONE IS THE LARGE V URSINUS (ALTHOGH STRANGELY, THEY HAVE THE SAME HEAD LENGTH AND DON'T SEEM TO HAVE ANY DIFFERENT LANDMARK ORDERS OF MAGNITUDE!), TRICHOSURUS ALSO LOOKS BIG AND SHOULD BE CHECKED; AEPYPRYMNUS AT THE BOTTOM ALSO DOESN'T LOOK RIGHT. TARSIPES ROSTRATUS IS TOO LARGE. SUGGEST WAITING UNTIL ALL BODY MASSES ARE AVAILABLE.

```{r}


Csize_ordered <- GPA_AllSpecimens$Csize[order(GPA_AllSpecimens$Csize)]

plot( Csize_ordered ~ seq(1:length(Csize_ordered)))
text( Csize_ordered ~ seq(1:length(Csize_ordered)), label = names(GPA_AllSpecimens$Csize), cex = 0.5)


```


#Importation of partition map, ensure it's a factor; The landmark dataset will now be split into two partitions; basicranium and the rest of the skull. 

```{r}

### Import the  partition map, ensure it's a factor
part.gp=as.vector(read.csv("../Data/Raw/partitions.csv", header=FALSE))
part.gp=as.factor(part.gp$V1)


```


#Make a list of species as the basis for the classifier file
```{r}


#speclist <- matrix(unlist (strsplit(gsub(".txt", "", filelist), "_" )), ncol=3, byrow = TRUE )
#rownames(speclist) <- gsub(".txt", "", filelist); colnames(speclist) <- c("Genus", "Species", "Museum_ID")

#write.csv(speclist, file = "../Data/Raw/Species_classifier.csv")

speciesList <- read.csv("../Data/Raw/Species_classifier.csv")

```





#AVERAGING MULTIPLE SPECIMENS OF ONE SPECIES

```{r}

# shape data
ind.coords <- aggregate(two.d.array(GPA_AllSpecimens$coords) ~ dimnames(GPA_AllSpecimens$coords)[[3]], FUN=mean)[,-1]
rownames(ind.coords) <- unique(dimnames(GPA_AllSpecimens$coords)[[3]])
ind.coords <- arrayspecs(ind.coords, p=ncol(ind.coords)/3, k=3)
#ind.coords_gpa <- gpagen (ind.coords) #Leonie did this but double-sliding the configurations seems dubious to me.

# centroid size data
ind.Csize <- as.vector(aggregate(GPA_AllSpecimens$Csize ~ dimnames(GPA_AllSpecimens$coords)[[3]], FUN=mean)[,-1])
names(ind.Csize) <- unique(names(GPA_AllSpecimens$Csize))


# ind.Csize - These are the individuals Csize data for plotting specimen allometry graphs

#are dimnames identical i.e. will the aligned properly?
dimnames(GPA_AllSpecimens$coords)[[3]]==names(GPA_AllSpecimens$Csize)

dimnames(ind.coords)[[3]]==names(ind.Csize)

#Turning into gdf frame and giving names
GPA_Allmean <-geomorph.data.frame(coords=ind.coords, Csize=ind.Csize)

GPA_allmean_consensus <- GPA_AllSpecimens$consensus

dimnames(GPA_Allmean$coords)[[3]] <-dimnames(ind.coords)[[3]]
names(GPA_Allmean$Csize) <- names(ind.Csize)


```


#GPA for basicranial landmarks
```{r}
#Subset basicranial landmarks
Data_basi=Data[which(part.gp==2),,]
#double-check that the correct number of landmarks has been subtracted
attributes(Data_basi)

#GPA
GPA_basi <- gpagen(Data_basi, Proj = TRUE)


#AVERAGING MULTIPLE SPECIMENS OF ONE SPECIES

#Two of the species in this dataset have multiple specimens which means they must be averaged prior to analysis. Once this is performed, GPA must be run again to incorporate these changes 


# shape data
basi.coords <- aggregate(two.d.array(GPA_basi$coords) ~ dimnames(GPA_basi$coords)[[3]], FUN=mean)[,-1]
rownames(basi.coords) <- unique(dimnames(GPA_basi$coords)[[3]])
basi.coords <- arrayspecs(basi.coords, p=ncol(basi.coords)/3, k=3)
# centroid size data
basi.Csize <- as.vector(aggregate(GPA_basi$Csize ~ dimnames(GPA_basi$coords)[[3]], FUN=mean)[,-1])
names(basi.Csize) <- unique(names(GPA_basi$Csize))
# basi.Csize - These are the individuals Csize data for plotting specimen allometry graphs

#are dimnames identical i.e. will the aligned properly?
dimnames(GPA_basi$coords)[[3]]==names(GPA_basi$Csize)

dimnames(basi.coords)[[3]]==names(basi.Csize)

#Turning into gdf frame and giving names
GPA_Basimean <-geomorph.data.frame(coords=basi.coords, Csize=basi.Csize)
dimnames(GPA_Basimean$coords)[[3]] <-dimnames(basi.coords)[[3]]
names(GPA_Basimean$Csize) <- names(basi.Csize)

GPA_basimean_consensus <-GPA_basi$consensus

```



#GPA for the REST OF THE SKULL


```{r}

#Subset the rest of the skull's landmarks lms
Data_reskull=Data[which(part.gp==1),,]
#double-check that the correct number of landmarks has been subtracted
attributes(Data_reskull)

#GPA

GPA_reskull <- gpagen(Data_reskull,Proj = TRUE)


#AVERAGING MULTIPLE SPECIMENS OF ONE SPECIES

#Two of the species in this dataset have multiple specimens which means they must be averaged prior to analysis. Once this is performed, GPA must be run again to incorporate these changes 

#Remember= dim() is the dimensions of an object

# shape data
ind.coords <- aggregate(two.d.array(GPA_reskull$coords) ~ dimnames(GPA_reskull$coords)[[3]], FUN=mean)[,-1]
rownames(ind.coords) <- unique(dimnames(GPA_reskull$coords)[[3]])
ind.coords <- arrayspecs(ind.coords, p=ncol(ind.coords)/3, k=3)
# centroid size data
ind.Csize <- as.vector(aggregate(GPA_reskull$Csize ~ dimnames(GPA_reskull$coords)[[3]], FUN=mean)[,-1])
names(ind.Csize) <- unique(names(GPA_reskull$Csize))
# ind.Csize - These are the individuals Csize data for plotting specimen allometry graphs

#are dimnames identical i.e. will the aligned properly?
dimnames(GPA_reskull$coords)[[3]]==names(GPA_reskull$Csize)

dimnames(ind.coords)[[3]]==names(ind.Csize)

#Turning into gdf frame and giving names
GPA_Restmean <-geomorph.data.frame(coords=ind.coords, Csize=ind.Csize)
dimnames(GPA_Restmean$coords)[[3]] <-dimnames(ind.coords)[[3]]
names(GPA_Restmean$Csize) <- names(ind.Csize)

GPA_restmean_consensus <- GPA_reskull$consensus

```

#Making coordinate datasets with size correction


```{r, echo=FALSE}


#THE BELOW USES THE COORDINATES OF THE DATASET WITH SPECIES AVERAGED; CSIZE WITH SPECIES AVERAGED;AND CONSENSUS CONFIG OF THE COORDINAGES WITH THE SPECIES AVERAGED.

#Make pgls of coords with size to get residuals (from the resulting pgls summary)
allom_all<-procD.pgls(coords~Csize, tree_synch, data=GPA_Allmean, iter=999)
summary(allom_all)
#Turn the residuals into an appropriate array
allom_res_all <-arrayspecs(allom_all$pgls.residuals,dim(GPA_Allmean$coords)[[1]],3)
#Add the consensus shape (from the GPA) to residuals if you want to understand the landmark variation in the residuals
Allom_all_plus_consensus <- allom_res_all + array(GPA_allmean_consensus, dim(allom_res_all))


allom_basi<-procD.pgls(coords~Csize, tree_synch, data=GPA_Basimean, iter=999)
summary(allom_basi)
allom_res_basi <-arrayspecs(allom_basi$pgls.residuals,dim(GPA_Basimean$coords)[[1]],3)
Allom_basi_plus_consensus <- allom_res_basi + array(GPA_basimean_consensus, dim(allom_res_basi))

allom_rest<-procD.pgls(coords~Csize, tree_synch, data=GPA_Restmean, iter=999)
summary(allom_rest)
allom_res_rest <-arrayspecs(allom_rest$pgls.residuals,dim(GPA_Restmean$coords)[[1]],3)
Allom_rest_plus_consensus <- allom_res_rest + array(GPA_restmean_consensus, dim(allom_res_rest))




open3d()
plot3d(Allom_all_plus_consensus[,,10], asp=FALSE)
text3d(Allom_all_plus_consensus[,,10], texts=c(1:58))


```


#To understand the main variation in shape between marsupial species I run a Principal Component Analysis (PCA) on the coordinate data. Using the function gm.prcomp from the geomorph package, the resulting graph will provide the two PC axes that explain the most variation between species coordinates. To understand the variance of each PC axis, I will run the function pc.sumary$importance to view the eigenvalues. These eigenvalues aid in further analyses.

#Each PC represents a particular aspect of the overall shape variation, see more in the chunk

```{r}

#Run PCA for all three partitions
PCA_mean <- gm.prcomp(GPA_Allmean$coords)

PCA_basi <- gm.prcomp(GPA_Basimean$coords)

PCA_rest <- gm.prcomp(GPA_Restmean$coords)


```



#XXXredo with new mean spec

```{r}
#mean specimen? 
findMeanSpec(GPA_Allmean$coords)

#Make mean shape that the points can be plotted on
MeanShape=mshape(GPA_Allmean$coords)
  
  #Or, for 3d plotting, call up the coordinates of the mean specimen; remember that the radius of the spheres might need to increase because this is not GPA'ed
RefCoords <- Data[,,7]

Dviv <- read.ply("../Data/Dasyurus_viverrinus_ascii_small.ply")


Meanwarp_Dviv <- warpRefMesh(Dviv,RefCoords*1000, MeanShape )


```


#Package into RDA file so it's available in the analyses
```{r}

save(Data,GPA_AllSpecimens,GPA_basi, GPA_reskull, GPA_Allmean, GPA_Allmean,GPA_Basimean,GPA_Restmean, Allom_all_plus_consensus,Allom_basi_plus_consensus,Allom_rest_plus_consensus,allom_res_all,allom_res_basi, allom_res_rest, speclist,part.gp, tree_synch, PCA_mean, PCA_basi, PCA_rest, MeanShape, RefCoords,  file = "../Data/Processed/processed_data_for_mars_phylo_shape.rda")


```
