==> Analyses/02_1_Clustering code.Rmd <==
---
title: "Clustering code"
author: "Vera Weisbecker, code supplied partially by Pietro Viacava"
date: "27 June 2019"
output: html_document
---

```{r}
library(rgl)
library(geomorph)
library(ape)
library(plyr)
library(geiger)
library(caper)
library(phytools)
library(nlme)
library(Morpho)
library(ClassDiscovery)# for colourful hclust plots
library(phangorn)


#library(landvR)

# My rmd file for some reason won't recognise it s local file, so this sets the wd to local
library(rstudioapi)
setwd(dirname(getActiveDocumentContext()$path))

load (file = "../Data/Processed/processed_data_for_mars_phylo_shape.rda")


```

#just some draft code for the tree stuff - the tasks are:
# 1) Find the youngest node
# 2) prune the gree so that the youngest node is now a terminal taxon
# 3) use the ancestral shape at that node as "species" 
# 4) make new random-to-molecular treedist comparison distribution
# 5) compare new morph-to-molecular treedist. Probably use Bhattacharya coefficient to assess if different from random dist
# 6) repeat
# 7) when this algorithm has "eaten" its way down the tree, visualise. In particular, I expect that most phylogenetic signal comes from higher, not lower branches but it would be super valuable to know if this is the case.


```{r}

#1) find the youngest node

tree_synch$edge.length

which(tree_synch$edge.length == max(tree_synch$edge.length))

plot(tree_synch)
edgelabels(tree_synch$edge.length)

```





```{r, echo=TRUE}

Partitions <- list( GPA_Allmean = GPA_Allmean$coords, GPA_Basimean = GPA_Basimean$coords, GPA_Restmean = GPA_Restmean$coords, Allom_all = Allom_all_plus_consensus, Allom_basi = Allom_basi_plus_consensus, Allom_rest = Allom_rest_plus_consensus)

Euclidean_list <- list()
UPGMA_trees <- list(paste("UPGMA_tree_", names(Partitions)))
NJ_trees <- list(paste("NJ_tree_", names(Partitions)))

#compute euclidean distances

for (i in (1:length(Partitions))){ 
  

  Euclidean_list [[i]] <- dist(two.d.array(Partitions[[i]]), method = "euclidean")
  
  UPGMA_trees [[i]] <-  upgma(Euclidean_list[[i]])
  
  NJ_trees [[i]]    <- NJ(Euclidean_list[[i]])
  
    }

#The below is required to get the Morpho_trees object into a multiPhylo object for later allowing treedist to run. This tip is from Liam Revell's blog (http://blog.phytools.org/2016/03/trick-to-append-object-of-class-phylo.html)

Morpho_trees <- UPGMA_trees[[1]]

class(Morpho_trees) <- "multiPhylo"
  
Morpho_trees <- c(UPGMA_trees, NJ_trees)
names(Morpho_trees) <- c(paste("UPGMA_tree_", names(Partitions), sep= ""),
                         paste("NJ_tree_", names(Partitions), sep = "" ))



```


#Tree congruence of the NJ and UPGMA trees relative to the "true" molecular phylogeny. Easy!


```{r}



tree_distances_morph <- list()

for (i in (1:length(Morpho_trees))){
  
  tree_distances_morph [[i]] <- treedist(Morpho_trees[[i]], tree_synch, check.labels = TRUE)
  
}

names(tree_distances_morph) <- names(Morpho_trees)

```

#now make comparison of random trees with "true" tree

```{r}

tree_distances_random <- list()

Random_trees <- rmtree(10000, length(tree_synch$tip.label), rooted = TRUE, tip.label = tree_synch$tip.label, br = runif)

for (i in (1:length(Random_trees))){
  
  tree_distances_random [[i]] <- treedist(Random_trees[[i]], tree_synch, check.labels = TRUE)
  
}

```

```{r}

#Make matrices of distance results for testing. I feel a bit silly with this behemoth of code but I don't know how to improve on this to get it into the data frame format it requires for plotting

Diffs_random <- sapply(tree_distances_random, function(x) c (x[1], x[2], x[3], x[4]), simplify = TRUE);
Diffs_random <-t(Diffs_random); names(Diffs_random) <- names(tree_distances_morph)[[1]]; Diffs_random <- as.data.frame(Diffs_random)

Diffs_morph <- sapply(tree_distances_morph, function(x) c (x[1], x[2], x[3], x[4]), simplify = TRUE);
Diffs_morph <-t(Diffs_morph); names(Diffs_morph) <- names(tree_distances_morph)[[1]]; Diffs_morph <- as.data.frame(Diffs_morph)


```

#Make four plots each with the random and "real" trees

```{r}
#produce density distribution

dens_symm_diff_random <- density (Diffs_random$symmetric.difference )
dens_brscore_diff_random <- density (Diffs_random$branch.score.difference )
dens_pathdiff_random <- density(Diffs_random$path.difference)
dens_quad_pathdiff_random <- density(Diffs_random$quadratic.path.difference)

Densities_random <- list( dens_symm_diff_random, dens_brscore_diff_random, dens_pathdiff_random, dens_quad_pathdiff_random  )


par(mfrow = c(2,2))

for (i in (1:ncol(Diffs_morph))){
  
  plot ( range(min(Diffs_morph[i])-3, max(Densities_random[[i]]$x)+10), range(0, max(Densities_random[[i]]$y)), pch=20, col="white", type = "n", xlab = colnames(Diffs_morph[i]),
     ylab = "Density", bty="n")
  
  polygon(Densities_random[[i]], col="plum1", border="plum1")
  arrows(x0 = mean(unlist(Diffs_morph[i])), x1 = mean(unlist(Diffs_morph[i])) , y0 = max(Densities_random[[i]]$y)/3, y1=0.001, col="lightslateblue", lwd = 2, angle = 15, length = 0.15)
 
  }

```


#The basicranium always performs worst! The whole dataset tends to do best

```{r}


par(mfrow = c(2,2))

for (i in (1:length (colnames(Diffs_morph)))){
  
  bar <- barplot(Diffs_morph[,i], col="lightslateblue", names.arg = colnames(Diffs_morph)[i],border = "purple" )
  
  text(bar, Diffs_morph[,i]- (Diffs_morph[,i]/2) , labels = rownames(Diffs_morph), srt =90, cex = 0.8 )
}


```




# Next: collapse shapes according to increasingly inclusive shapes (or use ancestral shapes?), then check for signal











##########################old and check code



#just plotting some stuff
```{r}


Dist_eucl_Allmean <- dist(two.d.array(GPA_Allmean$coords), method = "euclidean")

#hclust_upgma <- hclust(Dist_eucl_Allmean, method = "average")

#this isphangorn using hclust upgma. I checked, this really gives the same result as hclust 
upgma <- upgma(Dist_eucl_Allmean, method = "average")
nj <- NJ(Dist_eucl_Allmean)



plot.new()
par(mfrow=c(2,1))

plot(upgma)
plot(nj)


#Create object with all the morpho trees

```
















#Build phenograms based on Euclidean distances

```{r}

CVA_Clade <- CVA(GPA_Allmean$coords, groups = speclist$Clade, weighting = TRUE, plot = TRUE)

length(dimnames(GPA_Allmean$coords)[[3]])==length(speclist$Clade)

plot(CVAquollspops$CVscores[,1:2], bg=Data.pops, pch=21, typ="p", asp = 1)
legend('topright',legend=Data.pops,
       fill=Data.pops,cex=1)
title("Northern quoll cranial shape variation",cex.main=1.5)


Euclidist_quollspops <- CVAquollspops$Dist$GroupdistEuclid

  
the.cluster <- hclust(as.dist(Euclidist_quollspops))
plot(the.cluster)
plot(nj(Euclidist_quolls),main="Neighbor Joining") 
plot(upgma(Euclidist_quolls),main="UPGMA")

#Mahalanobis distances are not that good for building phenograms
Mahadist_quolls <- CVAquollspops$Dist$GroupdistMaha
plot(hclust(as.dist(Mahadist_quolls)))
plot(nj(Mahadist_quolls),main="Neighbor Joining") 
plot(upgma(Mahadist_quolls),main="UPGMA")
```

#Or you can also obtain a matrix of Procrustes distances from gpagen:
```{r}
#If I want to retreive the Procrustes distance matrix just like Gpagen does, but after averaging specimen configurations for multiple specimens, the below does this. 
GPA_Allmean_coordmatrix <-arrayspecs(GPA_Allmean$coords,"matrix")
ProcDists <- dist(GPA_Allmean_coordmatrix, method = "euclidean")

#This is for all specimens including duplicates

speclistAll<-read.delim("../Data/Raw/Mars_classifier_list_allspecs.txt")


#colsClade <- rainbow(n=3+length(levels(speclistAll$Clade)))
#or custom-code the colours:
levels(speclistAll$Clade)
colsClade=c("dark orange", "purple", "green", "red", "blue", "hotpink", "turquoise", "black", "forestgreen")





#Make shorter labels
Specmat <- matrix(unlist(strsplit(as.vector(speclistAll$Specimen), "_")),ncol=2,byrow = TRUE)
Genus <- substr(Specmat,1,3)
ClustLab <- paste(Genus[,1], Genus[,2], sep="_")
ClustLab <- as.vector(ClustLab)

#Make hc clusters of all three partitions
hclust_Allspecs <- hclust(as.dist(GPA_AllSpecimens$procD))
hclust_Basi <- hclust(as.dist(GPA_basi$procD))
hclust_Rest <- hclust(as.dist(GPA_reskull$procD))

par(mfrow = c(2,1))
#plotColoredClusters(hclust_Allspecs, labs=ClustLab , cols= colsClade[speclistAll$Clade], axes=FALSE, ann=FALSE, main= "Whole" )
plotColoredClusters(hclust_Basi, labs=ClustLab , cols= colsClade[speclistAll$Clade], axes=FALSE, ann=FALSE, main= "Basicranium", line=0); legend("topright", "Basicranium", bty="n")
plotColoredClusters(hclust_Rest, labs=ClustLab , cols= colsClade[speclistAll$Clade], axes=FALSE, ann=FALSE, main = "Rest", line=0 ); legend("topright", "Rest of Skull", bty="n")


```

#aaand this is the code I wrote with the bootstrapping included and saving each phenogram into a list:

```{r}
#PILOT phenogram bootstrapping

list_thephylo.cluster <- list()


for(replicate in 1:1000){
  
  sampledquolls<-sample(1:nrow(quollsdata), size = nrow(quollsdata), replace=TRUE)
  
  
  gpaspecs<-gpagen(A_reorder [,,sampledquolls], curves = NULL, surfaces = NULL, PrinAxes = TRUE,
                   max.iter = NULL, ProcD = TRUE, Proj = TRUE, print.progress = TRUE)
  
  All.shape<-gpaspecs$coords
  #Extract similarity matrix
  
  ProcD.mat <- gpaspecs$procD
  
  #Build phenogram
  
  the.cluster <- hclust(as.dist(ProcD.mat))
  
  #Convert phenogram to phylo format
  
  list_thephylo.cluster[[replicate]] <- as.phylo(the.cluster)
}



list_thephylo.cluster
